refactor: Extract common GutenbergKit editor configuration builder
Mitigate cryptic configuration bugs by reducing duplication.
